reorganize tests for entity config poller
